backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "RAG Api"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.3",
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "backoff==2.2.1",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "construct==2.5.3",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "einops==0.8.0",
    "email_validator==2.2.0",
    "fastapi==0.115.0",
    "fastapi-cli==0.0.5",
    "filelock==3.16.1",
    "fsspec==2024.9.0",
    "h11==0.14.0",
    "haystack-ai==2.6.1",
    "haystack-experimental==0.2.0",
    "httpcore==1.0.6",
    "httptools==0.6.2",
    "httpx==0.27.2",
    "huggingface-hub==0.25.2",
    "idna==3.10",
    "Jinja2==3.1.4",
    "jiter==0.6.1",
    "joblib==1.4.2",
    "lazy-imports==0.3.1",
    "Mako==1.3.5",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.1",
    "mdurl==0.1.2",
    "monotonic==1.6",
    "more-itertools==10.5.0",
    "mpmath==1.3.0",
    "networkx==3.4",
    "numpy==1.26.4",
    "openai==1.51.2",
    "packaging==24.1",
    "pandas==2.2.3",
    "pefile==2024.8.26",
    "pgvector==0.3.5",
    "pillow==10.4.0",
    "posthog==3.7.0",
    "psycopg==3.2.3",
    "psycopg-binary==3.2.3",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "pydantic_core==2.23.4",
    "Pygments==2.18.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "python-ptrace==0.9.9",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "regex==2024.9.11",
    "requests==2.32.3",
    "rich==13.9.2",
    "safetensors==0.4.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "sentence-transformers==3.2.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.35",
    "sqlmodel==0.0.22",
    "starlette==0.38.6",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "threadpoolctl==3.5.0",
    "tokenizers==0.20.1",
    "torch==2.4.1",
    "tqdm==4.66.5",
    "transformers==4.45.2",
    "typer==0.12.5",
    "typing_extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "uvicorn==0.31.1",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "websockets==13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "mypy>=1.12.1",
    "ruff>=0.7.0",
    "types-passlib>=1.7.7.20240819",
]

[tool.mypy]
strict = true
ignore_missing_imports = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py311"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
